Make new file for useStyles
