+++ client/src/pages/ProfileListings/ProfileListings.tsx
@@ -1,9 +1,6 @@
 import { useState, useEffect } from 'react';
 import Grid from '@material-ui/core/Grid';
-import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
-import CardHeader from '@material-ui/core/CardHeader';
-import CardActions from '@material-ui/core/CardActions';
 import CardActionArea from '@material-ui/core/CardActionArea';
 import CardContent from '@material-ui/core/CardContent';
 import Avatar from '@material-ui/core/Avatar';
@@ -14,16 +11,13 @@ import Divider from '@material-ui/core/Divider';
 import Rating from '@material-ui/lab/Rating';
 import Snackbar from '@material-ui/core/Snackbar';
 import MuiAlert, { AlertProps } from '@material-ui/lab/Alert';
-import Popover from '@material-ui/core/Popover';
-import IconButton from '@material-ui/core/IconButton';
 import SearchIcon from '@material-ui/icons/Search';
 import DateRangeIcon from '@material-ui/icons/DateRange';
 import CloseIcon from '@material-ui/icons/Close';
 import RoomIcon from '@material-ui/icons/Room';
-import { MuiPickersUtilsProvider, KeyboardDatePicker } from '@material-ui/pickers';
-import DateFnsUtils from '@date-io/date-fns';
 
 import DateSelectPopover from './DateSelectPopover';
+import useStyles from './useStyles';
 
 // Temporary user data to show functionality
 import { User, users } from './dummyUserData';
@@ -31,106 +25,6 @@ import { User, users } from './dummyUserData';
 interface MediaCardProps {
   user: User;
 }
-
-const useStyles = makeStyles({
-  root: {
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  title: {
-    fontWeight: 'bold',
-    margin: '2rem',
-  },
-  card: {
-    margin: '2rem',
-    width: '300px',
-  },
-  cardContentUpper: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    height: '250px',
-  },
-  cardContentLower: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-around',
-  },
-  cardAvatar: {
-    height: '100px',
-    width: '100px',
-  },
-  profilesContainer: {
-    alignSelf: 'center',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  searchDateContainer: {
-    flexDirection: 'row',
-    border: '1.25px solid lightgrey',
-    borderRadius: '5px',
-    justifyContent: 'center',
-    padding: '.5rem',
-  },
-  search: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  searchIcon: {
-    padding: '.25rem',
-    color: 'red',
-  },
-  roomIcon: {
-    color: 'red',
-    paddingRight: '.25rem',
-  },
-  cityContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-  },
-  popover: {
-    backgroundColor: 'rgba(0, 0, 0, .5)',
-  },
-  datePickerCard: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: '1rem',
-  },
-  datePickerHeader: {
-    alignSelf: 'flex-end',
-  },
-  datePickerTitle: {
-    paddingBottom: '1rem',
-  },
-  datePickerActions: {
-    flexDirection: 'column',
-  },
-  datePickerButton: {
-    marginTop: '1rem',
-    width: '7rem',
-  },
-  datePicker: {
-    marginRight: '5rem',
-    marginLeft: '5rem',
-    marginTop: '1rem',
-    marginBottom: '1rem',
-  },
-  dateRangeContainer: {
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  dateRangeIcon: {
-    paddingLeft: '1rem',
-    paddingRight: '1rem',
-    color: 'lightgrey',
-  },
-});
 
 function Alert(props: AlertProps) {
   return <MuiAlert elevation={6} variant="filled" {...props} />;
